+++ source/did-issuer-admin/frontend/src/pages/vc-management/issue-profile-management/IssueProfileManagementPage.tsx
@@ -79,7 +79,7 @@ const IssueProfileManagementPage = (props: Props) => {
     fetchIssueProfile(paginationModel.page, paginationModel.pageSize, null, null)
       .then((response) => {
         setRows(response.data.content);
-        setTotalRows(response.data.totalElements);
+        setTotalRows(response.data.total);
       })
       .catch((error) => {
         console.error("Failed to retrieve Issue Profiles. ", error);