+++ CHANGELOG.md
@@ -1,5 +1,24 @@
 # Changelog
 
+## [Unreleased] (v2.0.0)
+
+## v2.0.0 (2025-05-30)
+- Admin Console
+    - ZKP Management
+    - Adjusted some menu positions
+- Added functionality for Issuing ZKP Credential
+- Support for Besu blockchain
+
+### 🚀 New Features
+- Admin Console
+    - ZKP Management
+    - Credential Schema Manage
+    - Credential Definition Manage
+- Removed hardcoded logic and sample data from Issuer Server through Admin Console
+  - Removed hardcoded YAML configuration values
+  - Removed sample data previously embedded in Java source code
+
+
 ## [Unreleased] (v1.0.1)
 
 ### 🚀 New Features